+++ zia/services/firewallpolicies/filteringrules/firewall_filtering_rule_test.go
@@ -88,7 +88,7 @@ func cleanResources() {
 	}
 }
 
-func TestDLPWebRule(t *testing.T) {
+func TestFirewallFilteringRule(t *testing.T) {
 	name := "tests-" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
 	updateName := "tests-" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)